refactor: move handlers to corresponding packages
refs #109
