+++ message_handler/handlers.py
@@ -1,0 +1,5 @@
+from enum import Enum
+
+
+class MessageHandlers(Enum):
+    RabbitMQ = "rabbitMQ",
